+++ src/str.c
@@ -6,8 +6,8 @@ sf_str sf_str_fmt(const char *format, ...) {
     va_list arglist;
 
     va_start(arglist, format);
-    const unsigned long long size =
-        (unsigned long long)vsnprintf(NULL, 0, format, arglist);
+    const size_t size =
+        (size_t)vsnprintf(NULL, 0, format, arglist);
     va_end(arglist);
 
     char *fmt = calloc(1, size + 1);
